Got eslint working again and fixed linter error/possible bug.
Get eslint working again with a basic configuration. Running it, it complained
about getTextToCompare(). There is no `Token.string`, only `Token.str`.
This looks like a real bug. Unfortunately, I failed to write a test
where this bug is actually tested. The underlying reason is that I do
not yet understand what an "atomic tag" is supposed to mean. If it is
e.g. an iframe, the comments seem to suggest that the whole iframe
should be compared. But this does not happen. In fact, there are tests
('iframe Differences' > 'should show two widgets are the same if their
data attributes are the same') that test the exact opposite.

So I will keep it as it is right now, and concentrate on getting more
eslint checks.